requirements.txt
aioquic==1.0.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Flask==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==2.1.5
mitmproxy==10.3.1
mitmproxy-macos==0.5.2
mitmproxy_rs==0.5.2
msgpack==1.0.8
outcome==1.3.0.post0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
protobuf==5.27.0
publicsuffix2==2.20191221
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pylsqpack==0.3.18
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyperclip==1.8.2
PySocks==1.7.1
pytest==8.2.2
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
selenium==4.22.0
service-identity==24.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
tornado==6.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
urwid-mitmproxy==2.1.2.1
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
zstandard==0.22.0
